maine_career_backend/requirements.txt
# Maine Career Wealth Navigator — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core ML ───────────────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
scipy>=1.12.0

# ── SBERT (required — no TF-IDF fallback) ────────────────────────────────────
# Downloads all-MiniLM-L6-v2 (~90 MB) on first pipeline run; cached locally.
sentence-transformers>=2.6.0

# ── Clustering ensemble ───────────────────────────────────────────────────────
hdbscan>=0.8.33          # Density-based; installs joblib, cython extensions

# ── Dimensionality reduction (Wealth Galaxy 2D map) ──────────────────────────
umap-learn>=0.5.6

# ── API layer ─────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ── Optional (future resume upload endpoint) ──────────────────────────────────
python-multipart>=0.0.9
